+++ src/docs/Docs.tsx
@@ -10,8 +10,11 @@ export function Docs() {
       style={{ display: `none` }}
       headerOptions={{ Exit: hideDocs }}
     >
-      <div className="flex justify-center bg-zinc-950 scrollbar-y select-text">
-        <div className="px-2 py-2 docs-content" dangerouslySetInnerHTML={{ __html: docs }}></div>
+      <div className="flex justify-center bg-zinc-950 scrollbar-y">
+        <div
+          className="px-2 py-2 select-text docs-content"
+          dangerouslySetInnerHTML={{ __html: docs }}
+        ></div>
       </div>
     </Window>
   );